Move ToolDescriptionHelpers into Model\Tools
