+++ pwiz_tools/Skyline/Skyline.csproj
@@ -414,6 +414,7 @@
       <DependentUpon>TutorialTextResources.resx</DependentUpon>
     </Compile>
     <Compile Include="Controls\TargetColumn.cs" />
+    <Compile Include="Model\Tools\ToolDescriptionHelpers.cs" />
     <Compile Include="ToolsUI\ToolService.cs" />
     <Compile Include="EditUI\EditLinkedPeptidesDlg.cs">
       <SubType>Form</SubType>
@@ -5802,7 +5803,6 @@
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Controls\Tools\" />
-    <Folder Include="Model\Tools\" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Target Name="AfterPublish">
