+++ pwiz_tools/Skyline/Test/ToolDescriptionTest.cs
@@ -21,6 +21,7 @@ using System.IO;
 using System.Linq;
 using System.Windows.Forms;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using pwiz.Skyline.Model.Tools;
 using pwiz.Skyline.ToolsUI;
 using pwiz.SkylineTestUtil;
