+++ pwiz_tools/Skyline/ToolsUI/ToolMacros.cs
@@ -26,6 +26,7 @@ using System.Xml.Schema;
 using System.Xml.Serialization;
 using pwiz.Common.SystemUtil;
 using pwiz.Skyline.Model;
+using pwiz.Skyline.Model.Tools;
 using pwiz.Skyline.Properties;
 using pwiz.Skyline.Util;
 using pwiz.Skyline.Util.Extensions;
